requirements.txt
crc==4.1.0
paho-mqtt==1.6.1
pydantic==1.10.5
pyserial==3.5
typing_extensions==4.5.0
